requirements.txt
langchain>=0.1.0
sentence-transformers>=2.2.2
qdrant-client>=1.7.0
torch>=2.0.0
transformers>=4.30.0
python-dotenv>=1.0.0
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.23
mysqlclient>=2.2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
requests>=2.31.0
beautifulsoup4>=4.12.2
pydantic>=2.5.2
python-docx>=0.8.11
PyPDF2>=3.0.0
openai>=1.3.5 